requirements.txt
imagecodecs==2023.9.18
keras==2.10.0
Keras-Preprocessing==1.1.2
matplotlib==3.8.1
numpy==1.26.2
pandas==2.1.3
scikit-image==0.22.0
scikit-learn==1.3.2
scipy==1.11.3
tensorflow-cpu==2.10.0
tifffile==2023.9.26